simple_calculator.py

+++ simple_calculator.py
@@ -1,7 +1,7 @@
 # Simple Calculator
 
-num_1 = int(input("Enter the first number: "))
-num_2 = int(input("Enter the second number: "))
+num_1 = float(input("Enter the first number: "))
+num_2 = float(input("Enter the second number: "))
 op = input("Enter the operation (+,-,*,/): ")
 
 if op == "+":
